Rename component file to its name
